+++ demo-projects/blog/package.json
@@ -14,7 +14,7 @@
     "start": "cross-env NODE_ENV=production keystone start"
   },
   "dependencies": {
-    "@apollo/react-hooks": "^3.1.5",
+    "@apollo/react-hooks": "^4.0.0",
     "@apollo/react-ssr": "^3.1.5",
     "@arch-ui/layout": "^0.2.13",
     "@arch-ui/typography": "^0.0.17",
@@ -40,10 +40,10 @@
     "date-fns": "^2.14.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "graphql-tag": "^2.10.3",
+    "graphql-tag": "^2.10.4",
     "isomorphic-unfetch": "^3.0.0",
     "next": "^9.4.4",
-    "node-fetch": "^2.3.0",
+    "node-fetch": "^2.6.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   },
