+++ demo-projects/meetup/package.json
@@ -14,7 +14,7 @@
     "start": "cross-env NODE_ENV=production keystone start"
   },
   "dependencies": {
-    "@apollo/react-hooks": "^3.1.5",
+    "@apollo/react-hooks": "^4.0.0",
     "@apollo/react-ssr": "^3.1.5",
     "@emotion/core": "^10.0.28",
     "@keystonejs/adapter-mongoose": "^9.0.0",
@@ -39,7 +39,7 @@
     "express": "^4.17.1",
     "facepaint": "^1.2.1",
     "get-contrast": "^2.0.0",
-    "graphql-tag": "^2.10.3",
+    "graphql-tag": "^2.10.4",
     "isomorphic-unfetch": "^3.0.0",
     "lodash.uniqby": "^4.7.0",
     "next": "^9.4.4",
