+++ packages/apollo-helpers/package.json
@@ -18,7 +18,7 @@
     "apollo-cache": "^1.3.5",
     "apollo-link": "^1.2.14",
     "apollo-utilities": "^1.3.4",
-    "graphql-tag": "^2.10.3",
+    "graphql-tag": "^2.10.4",
     "hoist-non-react-statics": "^3.0.1",
     "lodash.mapvalues": "^4.6.0",
     "memoize-one": "^5.1.1",
