+++ packages/app-graphql-playground/package.json
@@ -16,7 +16,7 @@
     "body-parser": "^1.18.2",
     "chalk": "^2.4.2",
     "falsey": "^1.0.0",
-    "graphql-tag": "^2.10.3",
+    "graphql-tag": "^2.10.4",
     "object-hash": "^1.3.1",
     "terminal-link": "^1.3.0"
   },
