+++ packages/fields-cloudinary-image/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@keystonejs/fields-cloudinary-image",
+  "description": "KeystoneJS Cloudinary Image Field Type",
+  "version": "1.0.0",
+  "author": "The KeystoneJS Development Team",
+  "license": "MIT",
+  "repository": "https://github.com/keystonejs/keystone/tree/master/packages/fields-cloudinary-image",
+  "homepage": "https://github.com/keystonejs/keystone",
+  "engines": {
+    "node": ">=10.0.0"
+  },
+  "dependencies": {
+    "@arch-ui/button": "^0.0.20",
+    "@arch-ui/fields": "^3.0.3",
+    "@arch-ui/input": "^0.1.10",
+    "@arch-ui/layout": "^0.2.13",
+    "@arch-ui/lozenge": "^0.0.16",
+    "@arch-ui/theme": "^0.0.10",
+    "@babel/runtime": "^7.8.4",
+    "@emotion/core": "^10.0.28",
+    "@iframely/embed.js": "^1.1.1",
+    "@keystonejs/adapter-knex": "^11.0.0",
+    "@keystonejs/adapter-mongoose": "^9.0.0",
+    "@keystonejs/build-field-types": "^5.2.10",
+    "@keystonejs/field-content": "^7.0.2",
+    "@keystonejs/fields": "^14.0.0",
+    "@primer/octicons-react": "^10.0.0",
+    "image-extensions": "^1.1.0",
+    "node-fetch": "^2.6.0",
+    "pluralize": "^7.0.0",
+    "prop-types": "^15.7.2",
+    "query-string": "^6.8.1",
+    "slate": "^0.58.4",
+    "slate-drop-or-paste-images": "^0.9.1",
+    "slate-react": "^0.58.4"
+  },
+  "peerDependencies": {
+    "react": "^16.13.1"
+  },
+  "devDependencies": {
+    "react": "^16.13.1"
+  },
+  "main": "dist/fields-cloudinary-image.cjs.js",
+  "module": "dist/fields-cloudinary-image.esm.js"
+}
