+++ packages/fields-cloudinary-image/views/Cell.js
