+++ packages/fields-cloudinary-image/views/blocks/single-image.js
