+++ packages/fields-oembed/OEmbedBlock.js
@@ -2,8 +2,8 @@ import pluralize from 'pluralize';
 import { importView } from '@keystonejs/build-field-types';
 
 import { Block } from '@keystonejs/field-content/Block';
-import OEmbed from './';
-import RelationshipType from '../Relationship';
+import OEmbed from '.';
+import RelationshipType from '../fields/src/types/Relationship';
 
 const RelationshipWrapper = {
   ...RelationshipType,
