+++ packages/keystone/package.json
@@ -30,7 +30,7 @@
     "fs-extra": "^7.0.0",
     "globby": "^11.0.0",
     "graphql": "^14.6.0",
-    "graphql-tag": "^2.10.3",
+    "graphql-tag": "^2.10.4",
     "graphql-type-json": "^0.3.1",
     "lodash.flattendeep": "^4.4.0",
     "micro-memoize": "^4.0.9",
