+++ packages/oembed-adapters/package.json
@@ -12,11 +12,7 @@
     "node": ">=10.0.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.8.4",
-    "@emotion/core": "^10.0.28",
-    "@iframely/embed.js": "^1.1.1",
-    "@keystonejs/build-field-types": "^5.2.5",
-    "node-fetch": "^2.3.0"
+    "@keystonejs/fields-oembed": "^1.0.0"
   },
   "peerDependencies": {
     "react": "^16.13.1"
